Added a prepareSurveyChange notifier to IOMan [devel]
Summary:
The problem:
IOM().surveyToBeChanged notifier had a callback to uiODApplMgr::surveyToBeChanged that triggered the NotSavedPrompter via uiODMain::askStore(). If the user cancelled the NotSavedPrompter, indicating he wants to cancel the entire survey change pipeline, the other CallBacks in the IOM().surveyToBeChanged notifier kept being called, tearing down parts of the application. This compromised the integrity of the current survey that the user does not want to change.

Solution:
Added an extra IOMan::prepareSurveyChange Notifier that should be used to prepare for a survey change (like saving unsaved objects) without tearing down anything. The idea is to keep the current survey perfectly functional after these callbacks. Then the surveyToBeChanged should be used for some cleaning up and should be treated as a point of no return.

- Also delayed setting of rootdir_ until the IOMan::close() call in reInit().
- Added a validIdx check in PickSetMgr.

Test Plan: Tested different sequences of survey change, like cancelling the NotSavedPrompter without saving, saving the objects in the NotSavedPrompter. All sequences give the desired outcomes.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16205