Fix for crash while importing objects from another survey
Summary:
OD-4877.
The main problem was IOMan::setTempSurvey was calling setRootDir that led to a reInit and triggered all sorts of notifiers including surveyToBeChanged which involves a set of destructive actions. Now setTempSurvey and cancelTempSurvey have independent implementations that chnage the IOMan::rootdir_ but the effect is shallow and does not trigger any destructive actions across the application.

Test Plan: Imported attribute set and CBVS cube from another survey. The objects from the guest survey are listed and the integrity of the host survey is not affected afterwards.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16222