Fix for Simple file (ASCII) Seismic import
Summary:
OD-4899.
For 2D, this tool always needs both trcNr and XY (either from file or from user input). TrcNr should be the primary identifier for traces. The Coord should only be used for new geometries or for overwriting existing geometries. If the user chooses not to overwrite an existing geometry (prompted by Geom2DImpHandler), the SeisTrcInfo gets the Coord values from existing 2D geometry. Any traces outside the range of the existing geometry are discarded. It would be nice to give a feedback to user on the discarded traces, but that seems difficult in the current design.

Test Plan: Tested 2D and 3D simple file import with different options. All seems to work fine now.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16243