updated how we generate list for firewall rule [devel]
Summary:
The earlier implemenation of reading firewall rule was not working with current implementation.
Now we decalre two list which contain OD_FORREWALL_EXCEPTION_EXEC and OD_FIREWALL_EXCEP_DESC which contain rule name and their description repectively. There is thirrd variableOD_FIREWALL_TYPE which defines the type of process it is. Currently we have only OD and Python. Also updated the re-reading of the python process when the python path is not provided during the run

Test Plan: When no path is passed, a dialog is popped up and when feeding the correct parameters. Python environments are re-read. On clicking ok, success message is shown

Reviewers: raman, arnaud

Reviewed By: raman, arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D16255