Avoid reset of PythonAccess modules list [devel]
ServiceMgrBase::pythEnvChangedReq is triggered when a batch program is successfully finished,
however that batch program uses a custom python source, not an internal python source.
Needed to add some code to make it understand that these two sources may be equivalent if they
use the source root filepath.
Ideally the server service implementation for batch programs should override that base function,
but this cannot be done in this branch.

[Jira: OD-4909]