fixed a crash, batch processing [devel]
Summary:
WHen a process was terminated by using the close button (X), and same process was relaunched, there was a crash.
This was because the objectset was being emptied in wrong fashion and with next run there was issue with incrementing the iterator. This has been fixed now.
Also checking for valid index instead of < 0 check earlier placed

Test Plan: Tested and crash is fixed. Second process is live and opendtect does not crash

Reviewers: raman, arnaud

Reviewed By: raman

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D16271