Fixed layer model crossplot data extraction for integrated properties
Strat::LayerSequence::getSequencePart returns nothing when an empty depth-gate is provided,
as it should not return layers of 0 thicknesses.
However we do need to have some layer to be able to extract a crossplot of integrated rock properties,
hence copying the first layer of the sequence if not empty, to avoid getting an empty datapointset.

[Jira: OD-4945]