Fixed incorrect datapack selection in the flatviewer
The booleans wva and vd in FlatView::Viewer::setPack are destinations,
thus not appropriate for selecting the current class members.
Fixes an issue in the layer modelling window, where the call to setPack
returns instead of effectively changing the datapack (as the wva datapack
is equal to the former vd datapack).

[Jira: OD-4957]