now adapted to set default translator [od]
Summary: With popping up the ML CC we can now set the defualt translator for writing the output. This selection can be overwritten with preffered translator

Test Plan:
Tested in 7.0 two cases :
laucnhed from OpendTect and default was set to CBVS. When selecting particular translator the same type of object was created
When laucnhed from Petrel, default was set to Petrel Direct. Upon chnaging the write output the same was reflected in the UI and object of selected type was created

Reviewers: raman

Reviewed By: raman

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16324