Added explicit units GUI to the sampled PDF creation dialog
Summary:
uiCreateDPSPDF must show the units when used in some contexts, to be saved in the created PDF.
Added a boolean in the constructor to specify that context.
uiPrDenFunVarSel now includes an optional uiUnitSel when required. Its initial state is set
either from the DataPointSet content (when not empty), or by using the guesss functions
already used by the PDF edition window.
The user must provide valid units in such context, in order to be able to save the created PDF.
This will ensure workflow that need them will get the required information.
+ general code cleanup, and workflow fixes: This dialog did not correctly select the properties
in the function setPrefDefNames (items were shifted, or duplicated)

[Jira: OD-4968]

Test Plan: discussed with a user, using the workflow described in the jira issue.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16328