Various improvements in the PDF edit dialog
Summary:
Sampled and Gaussian PDFs may have undefined units (was not stored up to and including v6.6).
Adding a label if at least one dimension does not have a unit of measure set, with a small
exception to handle PDFs created from seismic attributes, outside of well modelling workflows.
Also adding a label right of each unit of measure field, to show which one is guessed.
Switching the controls from Ok/Cancel to Save/Close.
When closing, if units were guessed, offering to save the PDF with the guessed units.

Also several cosmetic changes to the "View" option for sampled PDFs: Setting a single view range
from 0 to the max count (especially important for 3D PDFs), using a more standard colorbar,
switching to non-blocky type of display, starting the table and viewer in the middle of the dimension
ranges, and not at the edges (mostly zeros).

[Jira: OD-4970]

Test Plan: showed the various changes to the GUI to our users.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16337