Making the histogram slider interaction smoother
Summary: If the mouse moves out of range while sliding, we take the extreme value. This way, it's easier to move the sliders to the extremes.

Test Plan: Tested in the Fault Extraction window as well as the regular Histogram displays. The slider movement to the extremes is much smoother now.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16345