Fixed incorrect SI() while using the survey changer
The directories from IOM() were correct when using the survey changer class,
but SI() was not updated. Added the required push of the temporary SurveyInfo,
and corresponding pop when the survey changer is deleted.

Added an extensive test program that tests static functions accessed directly or
indirectly using SI(), in a standard test program that loops over multiple surveys.