Horizon Data values should be undefined if Z (geometry) is undefined
Summary:
If a Horizon is undefined at a BinID, the HorizonData values at that BinID should also be undefined.
Also removed some code duplication.

Test Plan: Removed a part of a horizon using a polygon. Then displayed a HorizonData in Basemap. The hole is visible in the HorizonData display as well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16390