Rework of Horiaon (2D/3D) Event Snapper
Summary:
The main task was to add an option to use the original Horizon if no events are found in the search gate for a particular position. But I ended up doing some cleanup as well.
1. Moved 3D snapper from Seis to EMAttrib in order to use EM::Horizon3D instead of a BinIDValueSet.
2. Got rid of the extra class for 2D, the Setup classes and the use of Hor2DSeisLineIterator.
3. Simplified the code wherever possible, removing old-fashioned stuff.

Test Plan: Snapped a constant Z horizon (2D and 3D) and tested both 'Erase' and 'Use original horizon' options. The results are as expected and the process is faster than before.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16395