UI now properly updated
Summary: Earlier  acall was being amde to update the unit list, this was missing a bool check to check whether the domain conversion is switched on or off. Part of OD-5025

Test Plan: Updating as it should. If domain is unchecked, the native survey domain units are displayed and if domain conversion is checked respective target domain units are displayed

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16414