solved issue in survey name update [devel]
Summary:
If someone updated active survey name, pressed okay on Survey Edit Parameter dialog. Then decided to close the the Survey Setup and Selection dialog by cross or Cancel instead of Okay.
The issue is the survey name in the list is updated, but next time when we open OpendTect it says specify the survey, since the name in the .od folder was not updated.
Now when the name is updated it checks if the current survey is the active survey by name comparison and then proceeds to update the survey file in .od folder. In this way even if the user change the name of active survey, they will be able to use the same survey next time they start OpendTect, OD-5052

Test Plan: Tested

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16428