Avoid unnecessary warning icons in front of non-attribute tree items
Summary:
This change avoids warning icons in front of attribute items that are no really attributes, and more importantly, that do not indicate a failure to load/compute. These include:
1. Default attribute item when a new Fault/FaultSet is added
2. Horizon tracking attributes and ZValues for a newly created Horizon

Test Plan: No warning signs are displayed in from of items mentioned in the summary. Also, warning signs are displayed in case of a genuine failure to load/compute.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16491