Fixes for VelocityDesc usage [devel]
Ensured the VelocityDesc::velunit_ is always set to the label of its corresponding UnitOfMeasure, made it private.
Fixed the saving of the default velocity model, but removing an obsolete API. This also means that a database conversion
is done the first time a uiVelDesc is constructed. The old key is kept to preserve readability of the database.
Fixed interaction issues between uiVelDesc and uiVelModelZAxisTransform,
added background scanning of the velocity ranges when a velocity model without scanned ranges is selected.