Fixed crash during Seis copy [devel]
The SeisSingleTraceProc was not ready in time, meaning that SeisCubeCopier could not
access its reader in time. That is now adjusted using a proper callback mechanism,
and as a result of that the class friendship is no longer required.

Also removed the trace cropping part: The reader already makes the component sub-selection,
thus it is no longer required to be done by SeisCubeCopier.