Removing usage of mToFeetFactor
Summary:
Double tranform mentioned in the issue is not reflected in the debug. Depth data is being used as it is in the transformed depth scene.
Since all the data is automatically stored in native depth format, no conversion should be required. Need to still check if someone overwrite this setting and the effect it has on the data

Test Plan: Tested

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16534