solved well log import issue [devel]
Summary: The label had space in it and that lead to wrong reading of names and unit. In turn it courrpted the imported data. Also the output location was not automatically set which is now being done

Test Plan: Tested by exporting object and improting them back. It works. I also changed the domain and imported them back correctly. The output location is also automatically read

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16543