Added a tool for ordering callbacks in a set [devel]
Summary

Sometimes it is required to use a specific order of callbacks,
added a function that allows changing the order of already attached callbacks in a set
Found that the implementation of ValVec::move was actually completely wrong, thus fixed it
and added the corresponding tests. Lucky this function was not used in any known code so far.
Test Plan

compiles and tested.