Fixed synchronized zooming with/without flattening in SynthRock
Summary:
several things did not go right: Mainly the getTime/getDepth
functions of TimeDepthModel were used by passing relative Z values from the viewers,
which was incorrect. These Z values have to be converted to absolute Z values before
the function call, and the returned values may only thereafter be converted to relative values.

Also made 'virtual' the call to setViewToBoundingBox. In case of coupled viewer, it should
only be applied to the active viewer, the other viewers will get updated automatically.

To achieve a good synchronization, it was essential to ensure that the flatviewer of the
depth model viewer (uiStratLayerModelDisp::vwr_) was reconstructed on model change before
the flatviewer of the synthetics viewer (uiStratSynthDisp::vwr_), this latter always
being the active viewer.

Finally there were problems setting up the view, the old implementation of uiFlatViewControl::getZoomOrPanRect
would not ever give a uiWorldRect centrered on the requested mouse position.

[Jira: OD-5162]

Test Plan:
Use the demo unconventional, synchronized zooming works now really well
with or without flattening. Checked also rubberband and zoom in/out of the standard
2D viewer, the returned view matches accurately the picked rectangular area.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16549