Fixed incorrect behaviour of the elastic properties selection GUI [devel]
The GUI was sized proportionally to the number of properties in the model.
This does not make sense, some formulas may have more inputs, others will
have less.
Now setting the number of input lines contextually, with a minimum of 8.
Also fixed many bugs in the interaction, saving or opening of
elastic properties selection in SynthRock