Removed silent error whent the proj database cannot be loaded
The CRS plugin initialization should fail when the proj database is missing,
or cannot be loaded.
The uiCRS plugin initialization should fail when the CRS plugin is not loaded successfully.

Forwarding the filename of the proj DB to the source files, and its full path so it can
be accessed in development environments without copying it to the binary directory.