Minor change in the cmake setup for CRS/proj [devel]
The SQlite library is a direct dependency of the proj library. Hence
the correct way to support it in cmake is to set the IMPORTED_LINK_DEPENDENT_LIBRARIES
target property for the PROJ::proj target, and not to make use of our own
OD_MODULE_EXTERNAL_RUNTIME_LIBS cmake variable.