Simplified the usage and installation of dependent targets [devel]
Using the cmake generator TARGET_RUNTIME_DLLS to copy all runtime DLLs of plugins
in the solution binary folder on Windows. This simplifies a lot the current cmake code.
The only consequence is that a qt.conf file needs to be generated as a post-build target,
with the full path to the msvc installation, to avoid having to copy the qt plugins and
other resources. This is setup for the Basic target.

A second custom command is also made to support pure runtime targets, like OpenSSL for the
Network module, and some Qt modules (Qt::QuickSupport, Qt::Svg).

The installation of these externals is also simplified accordingly.

Tested on Windows and Linux. No change in the final list of installed files. All plugins
load within the development environment.