Fix for CBVS files not deleted after removing entry from database
Summary: The default value of 'deep' is false. The argument deep has different meanings for different translators. For SEGYDirect it means deleting the linked SEGY files, while for CBVS it means deleting the main CBVS file (might still be a link). If the CBVS file is not a link, the regular delete message is displayed and implRemove is called with deep=false.

Test Plan: Deleted a regular seismic cube, the CBVS file is also deleted. For a lnked CBVS file, the special message is displayed and both options (delete cbvs' and 'delete link' work as expected.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16590