Reworked object selection mechanism [devel]
Summary:
Fixing issues where interfaces incorrectly report their ZDomain.
Both IOPar IOObjSelConstraints::require_ and IOObjSelConstraints::dontallow_ have historically
been used to constraint context selections, either by ZDomain::Def or type, however that is not
done correctly, and it will become between if all datasets start having a ZDomain key in their
database entry.
That selection was not read in the implementation of uiSeisSel::getZDomain, leading to a bug
in the time-to-depth conversion of seismic data.
Also in most cases IOObjSelConstraints::dontallow_ should not be used, as it only works correcly
when their a ternary choice: Choice A, Choice B, no choice. But there a more than two ZDomain::Def
used in the system, thus there must be a selection to specify exclusively the required ZDomain::Def,
exclude all others, but still include database entries without ZDomain tag, implicitely using SI().zDomain().

Test Plan:
GUI in time-to-depth of seismic is OK now, GUI in time-to-depth of horizons is OK for the input
but not (yet) for the output.
Still need to look into the code for usage of these two IOPars, and made adjustements as and where needed.

Reviewers: raman

Reviewed By: raman

Subscribers: prajjaval, nanne

Differential Revision: https://phabricator.dgbes.com/D16598