Added cmake support for plugin initialization [devel]
Summary:
Extending the scope of the macro OD_INSTALL_SIP to support the installation
of file to data/SurveyProviders, data/SeisTrcTranslators and data/FileSelProviders

Added the required packaging code for these last two.
Removed files that do not need to be version controlled.

Test Plan: all required files are created when configuring the project

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16617