Fixed issue in time-to-depth conversion GUI [devel]
The dataset selection was incorrect, because some uiSeisSel did not have the enabotherdomain_
flag set to true in their Setup. Fixed it by setting it to true where needed, and also
removed the call to requireZDomain for the interfaces which match the project ZDomain.

Ensured that the time/depth toggle in uiSeisSel is only displayed when more than one
ZDomain are allowed, and the allowed ZDomains must be time and depth.

Also removed usage of the unused uiSeisSel::Setup::zdomkey_ class member.