reloading the survey in everycase [devel]
Summary: Earlier just changing parameters did not reload the survey. Now relaoding the survey with change of parameters or name

Test Plan:
Tested by changing just the steps and also the name.
In both cases the survey was reloaded.

When any file was open a proper error came in case of survey rename
Kept survey file opened, in case of rename operation, proper error message was popped up.
In case of survey paramters changed, survey was reloaded without any error message

Picked new pickset in the scene, cahnged name and parameters and then saved the parameters, NSP prompt asked for saving of pickset and once the picksets were saved survey was relaoded with changed parameters Pickset were displayed in the 3DScene

Laucnehd a batch process, changed the name and parameters, in this case Batch program tracker was prompted and when all the processes were closed it allowed me to proceed and reloaded the survey with new parameters

Reviewers: raman

Reviewed By: raman

Subscribers: arnaud, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16653