fixed a crash in service IDs cleanup [devel]
Summary: Nothing should be removed in a loop that works on a increment. Objects should only be deleted in loop from back. This maintains the orders and ensures that requested indexes are actually deleted

Test Plan: Tesed, no crash seen. Crash earlier was when survey was renamed and a batch process was relaunched.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16662