Add replacement velocity to synthetic modelling
Summary:
See the description of the Jira issue:
https://dgbes.atlassian.net/browse/OD-5163
Implemented this improvement

Added a start depth and velocity to both Strat::LayerSequence and ElasticModel classes.
Adjusted the computation to have both parameters used with displaying a Strat::LayerSequence
or computing synthetic seismic using ElasticModel objects.
Reworked the GUI for the stochastic mode, both parameters are now located in a dedicated dialog.

[Jira: OD-5163]

Test Plan:
Used the demo unconventional project. Changes are obvious using the GUI.
Values are correctly read from and written in the simulation file and pseudowells.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16663