Fixed management of prestack 2D viewer [devel]
The pointer returned by uiViewer3DMgr::createMultiGather2DViewer and uiViewer3DMgr::create2DViewer
is not the pointer of the corresponding object, probably because of the inheritance.
As a result uiViewer3DMgr::viewer2DClosedCB was not able to find the index of the triggerd object
in uiViewer3DMgr::viewers2d_, resulting in a crash when closing od_main with a prestack 2D viewer still
on screen, because of double deletion (by Qt first, and then by uiViewer3DMgr::removeViewWin.

Was probably hidden on survey change by the implementation of uiViewer3DMgr::surveyToBeChangedCB.