Alternative solution for closing a uiMainWin
Summary:
This way of closing might be more controlled. With a 'Delete on Close' it's
added to a queue. When closing OpendTect, uiODMain might be deleted before
some of the windows that delete on close.

Test Plan:
Using the new debug statements in uiMainWin constructor and destructor,
checked that child windows are deleted before uiODMain.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16671