Added geometry context to prestack processing manager
Summary:
The PreStack::Processor, PreStack::ProcessManager and their GUI classes had no information on the context they were in: 3D processing, 2D processing or synthetic.
Added the required flow of information.

Also many classes around the prestack handling where still not using the TrcKey and TrcKeySampling classes correctly, setting the GeomID to 0 or worse. The uiSliceSel class is now eventually doing the expected work in the three domains.

Test Plan: compiles, GUI looks OK. Will need to check the actual processing (in batch and on-the-fly)

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16685