Better handling of survey creation from single 2D line
Summary:
If the line is parallel to one of the axes, two of the three coordinates for survey grid definition are identical and lead to a bogus survey definition. Now adding an extra inl (or crl) on each side gives a reasonable survey definition.
Also, when prompted to import the single line navigation file, the line name information is missing. Added a pop up to receive the line name from the user.

Test Plan: Created new survey from single 2D line as well as XY ranges for the line when the line is parallel to X or Y axis. The resulting survey is good. The imported line geometry gets the line name entered by the user.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16697