Fixed XPos values being set for layers in place of RelZ
Summary: I just fixed the wrong index. But we need a proper implementation of RelZ (we could vary it between 0 1and 1 in a Sequence, but a Layer knows nothing about the sequence). For lack of a better implementation, RelZ value is now set to layer depth.

Test Plan: Created a Stochastic model with RelZ being used in the formula for Density. The values are now taken from depth and not from XPos.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D16699