Implemented RelZ for use in layer modelling
Summary:
The layer generator gets its inputs from the Property::EvalOpts argument. The idea is to pass enough information along with this argument to facilitate the computation of RelZ, which is the relative depth of a layer (0-1) in its parent unit. Two different approaches are used for the following two cases:
1. When the parent unit has a fixed number of child layers: In this case the thickness of the parent unit is not known upfront, hence relz is computed based on the layer index (not on depth). See D16728. It still goes from 0 at the top of the parent unit to 1 at the bottom, and is a good approximation for lack of a better implementation.
2. When the thickness of the parent unit is fixed: In this case, SynthRock would pass the thickness of the parent unit as dummy EvalOpts::relz_ and the top depth of the parent unit as EvalOpts::refz_. When the actual layer is generated and its thickness is computed, this dummy relz_ is replaced with the actual relz_.

In future, when we have a more general implementation of RelZ, the newly added variable EvalOpts::refz_ can be used for passing a reference depth like the depth of a marker.

Test Plan: Created a new model using a complex stratigraphic framework. The value of RelZ appears to be correct in both the cases mentioned above. Also tested after saving and retriving the model.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D16727