Disabled edit of fault in transformed scene [devel]
Summary: Earlier faults were for edit, now edit can only be done in the non transformed scene

Test Plan: Working as it should, When the fault added to native scene, the editor is added at the bottom. When in the tranformed scene fault is only added for display. When toggling between these two faults, the editor panel is only displayed only when native scene fault is selected

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16733