Handling clipping and Histogram of data with many duplicates
Summary: If a dataset has many duplicate values (for instance an 8-bit dataset scaled and converted to floats), performance of the regular sorting algorithm is hit drastically. Added a function to detect such scenarios so that the special duplicate_sort can be used.

Test Plan: Displayed such data on sections and also displayed the histogram. The performance is at par with that of regular data.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16729