Allow selecting any seismic dataset in the pre-load manager
Summary:
The flag uiSeisSel::Setup::enabotherdomain was not set (to true)
for uiSeisPreLoadSel::seissel_, as a result the transformed datasets did not appear in the
list and could not be selected by the user.
However, when enabling this option, care must be taken for the (Z range) sub-selection:
A single uiSeisSubSel interface cannot be used accross different Z domains. As a result,
this revision adds the class uiMultiSeisSubSel, which can be used in such case. It makes
one uiSeisSubSel for each available ZDomain::Info loaded in the factory, and shows always
only one, corresponding to the dataset provided using the setInput function.

One should probably also use that interface in the SEG-Y import window.

[Jira: OD-5296]

Test Plan:
In a time project, could now pre-load time, depth, horizon-flattened and wheeler 3D datasets.
The volume subselection shows the expected Z domain label and range, limited to the dataset range.
The info section of the pre-load manager also shows the expected Z label, values and units.

2D to be tested (need to create transformed data first)

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16760