Added few tools to the HDF5 Link [devel]
Summary:
Added a rename function for groups and datasets.
Added API to retrieve the version of an HDF5 object
Added API to set and retrieve comments to groups and datasets.

Test Plan:
tested by expanding the test program.
Comments do NOT appear within the hdfview application, but can be cheched with
the h5dump application: h5dump -H -A 0 test.h5

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16768