Allow selecting any seismic dataset in the pre-load manager
Summary:
The flag uiSeisSel::Setup::enabotherdomain was not set (to true)
for uiSeisPreLoadSel::seissel_, as a result the transformed datasets did not appear in the
list and could not be selected by the user.
Since reworking the volume subselection GUI is too big in this branch, that interface is made
hidden for Z transformed datasets, i.e. the pre-loading will always use the entire dataset.

Also fixed label issues with transformed domains that do not have a unit of measure.

[Jira: OD-5296]

Test Plan:
In a time project, could now pre-load time, depth, horizon-flattened and wheeler 3D datasets.
The subselection field is hidden/shown where appropriate, and the pre-loaded dataset is useable
in the full/selected range respectively.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16771