Fixed log sampling issue
Summary:
when the first log has a crappy sampling, the default does not make sense

Q1: Another option is to leave it blank first, then fill it in based on the
used logs for new log calculation. uiWellLogCalc::fillSRFld does this when the
field is blank.

Q2: Do we need a survey default for this? To be stored in .defs

Q3: Do we want a combination of both? We can add a field where the user can
decide: 'survey default' or 'from used logs'.

Test Plan: Default sampling is now always 0.5 (ft) or 0.1524 (m)

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D16779