Updated root CMakeLists.txt for the Tut plugin exercise [devel]
Summary

Setting the minimum cmake version to v3.14. cmake warns the user when the version is less than 3.5,
and 3.14 is the first version of cmake that supports MSVC 2019, thus it is a good strict minimum
Also setting the default C++ language to c++17, to be inline with the SDK.

Test Plan

ran the programmer manual exercise. No more warning during the cmake configuration, both plugins
are build nicely without any warning. Libraries are placed within the ODWork project.