Added support for dox files in plugins
Summary:
crs.dox should not be placed at the root of the include folder, as CRS is not a core module.
Added general support for placing these files at the root of plugin directories, and cleaned up
slightly the cmake syntax.

Test Plan:
documentation is created. The CRS group and Coord namespace are correctly included, and functions
from the CRS plugin API are visible.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com/D16797