Added support for dox files in plugins
Summary

crs.dox should not be placed at the root of the include folder, as CRS is not a core module.
Added general support for placing these files at the root of plugin directories, and cleaned up
slightly the cmake syntax.
Added a dox file for the ODHDF5 plugin.

Test Plan

documentation is created. The CRS group and Coord namespace are correctly included, and functions
from the CRS plugin API are visible.