Fix for update issue in selection window after Velocity type change
Summary: After changing the velocity type, the IOObj::pars() is updated, the update was not carried over to the CtxtIOObj as the areEqual function only compares the MultiID (kind of misleading). In case of the the two IOObjs having the same MultiID, it is better to copy the rest of the data and still avoid deleting the old ioobj_.

Test Plan: Changed velocity type for a model from Vint to Vrms. The GUI is updated properly now. No adverse effect expected elsewhere.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16805