Updated the development environment creation [devel]
Added a skeleton to facilitate the distribution of plugins started from the Programmer's manual tutorial:
The root CMakeLists.txt file of the created development environment now features the following parameters:
- The product name, taken from the cmake project version name
- A vendor name
- A per-plugin version, package and license type specification

That information is used to:
1- Setup the static PluginInfo object for each plugin
2- Automatically create the package version file when installing the project

The ALO files are now disabled only for internal builds of OpendTect, but not when configuring the development environment.
Tested on Linux, to be tested on Windows later.