Fixed Well Copy issue where existing well's multiID would change
Summary: Source well is from the manager and it copy operation would change it's multiid. This was noticed in Well Table. In this revision I create a local refman<Well::Data> directly from the reader to act as a temporary object which is then copied into the new well

Test Plan: tested in well table

Reviewers: raman

Reviewed By: raman

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16855