Porting well copy fix from main
Summary:
The original revision was https://phabricator.dgbes.com/D16871 which had been accepted. When landing the change only one line had been landed, hence this revsion for rest of the code that needed to go, but somehow didn't.

Source well is from the manager and it copy operation would change it's multiid. This was noticed in Well Table. In this revision I create a local refman<Well::Data> directly from the reader to act as a temporary object which is then copied into the new well

Test Plan: tested in well table and Basic Well Manager

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16891