Fixed cmake configuration error with Qt6 on Linux [devel]
The target Qt6::DBus is now correctly marked by Qt as a dependency of Qt6::PrintSupport,
hence removing the need to manually add that target.
However it contains an sub-dependency to a non-existing target, 'dbus-1'. This incorrect
dependency must be removed while importing the Qt6::DBus target for the configuration to work.