Putting a warning icon in 3D Seismic Manager if obj is broken.
Summary: An isOK function has been added to SeisTranslator that is implemented in its derived classes. Similarly, SeisIOObjInfo too has an isOK function to access those in translators. if this function returns false, a warning pixmap is displayed in the manager alongside such entries.

Test Plan: tested in seismic manager

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16942