Reading Well Info before reading track [devel]
Summary:
The call to read well track was being made before reading well info which was not what we expect. Also the implementation in readReq funciton in wellman.cc ensures well info is available before going to read track.
This fix is also required for Petrel, since we are reading KB in info section and loss of this data leads to errorneous values

Test Plan: Tested, compiling and petrel data read properly

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16972