Fixed crash while adding units to a seemingly corrupt Stratigraphic framework
Summary: The crash in OD-5419 is caused by a Stratigraphic framework that was probably modified simultaneously from disfferent instalnces of OpendTect while SynthRock window was active. It is hard to tell exactly how it ended up in this state and how to prevent that in the code. Therefore, just fixed the crash by replacing a few unsafe casts with Dynamic cast followed by a nullptr check.

Test Plan: Added a unit to the reported framework successfully. No change in existing functionality expected.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com/D16982