solved issue of survey rename not possible
Summary:
OD-5426, the File::isInUse was returning a wrong value for a directory location and hence user was not able to rename the folder. Now a perrmsg has been placed in the function and the check has being removed.
Also the the msg for batch process running is no longer blocking. It is an error message on windows and normal message on linux and mac. If user wishes to return to the active survey that is allowed, but no changes in survey parameters allowed unless the batch process is closed on windows.

Test Plan:
following conditions were tested
1. Inactive survey started a batch process and picked pointsets, rename and change of parameter not allowed until batch process were closed and response to unsaved objects provided.
2. Started batch process in active survey, and changed parameters of different survey. Survey parameter changes allowed.
3. Started batch process in active survey, changed survey, and tried to change the parameter of survey from which batch process started. Rename operation not allowed, parameters allowed to be changed

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16994